scripts/requirements.txt
boto3>=1.34
pyyaml>=6.0
mcp>=1.0
